requirements.txt
gspread==5.3.2
python-dotenv==0.20.0
requests==2.25.1
schedule==1.1.0
